Changes to Molecule.makeHydrogensImplicit and makeHydrogensExplicit
If these functions are called, they run, no matter what the self.implicitHydrogens
value is (previously, if the molecule *thought* it was implicit, then a call to
makeHydrogensImplicit didn't do anything, even if there *was* an explicit H atom).

However, if no hydrogens are added or removed then, we don't reset the 
connectivity values.