Revert "Molecule.isIsomorphic always done with implicit hydrogens."
This reverts commit 6ff9e35dd7a37b562cf1976ecde83466a4adf328.

If you go through an implict/explicit cycle, then you end up with
the same number of hydrogen atoms, but they are different python objects,
so any pointers to them are lost. This includes pointers in a map,
which are used to stick labels on atoms. You therefore lose the ability
to label the atoms correctly when trying to apply a template to a
reactant structure, and the whole thing breaks.

If you leave them alone, then the atoms remain where they are and things work.