Better method for determining size of PES drawing.
The previous approach only checked the energies of the isomers, reactants,
and products, and did not check the transition state energies of the path
reactions. This new method first finds the minimum and maximum energy on
the surface (including stable configurations and transition states) and
uses these to compute the drawing height.

This was particularly noticeable when a transition state E0 was much
higher than the highest stable configuration E0. This doesn't happen
very often (unless the E0 is estimated poorly), but is certainly possible,
so we need to handle it properly.