KineticsFamily.getKinetics() now also returns isForward data for each result.
When getting the kinetics for a reaction generated from a template, the
kinetics could conceivably come in either direction if a match for the
reaction is found within a depository (e.g. training), as we don't force
these to be stored in the direction the template is defined in. Therefore
we need to mark the returned kinetics as being for the forward or reverse
direction, so they can be properly interpreted. All kinetics estimated
via group additivity or from the rules are by definition for the forward
direction, as before.

This fixes, among other things, the absurdly-high barrier for
CH3 + CH2OH -> CH3CH2OH seen in the PES for issue #44.