Don't get kinetics during KineticsFamily.generateReactions().
It turns out we can't unambiguously determine which direction (forward or
reverse) to get the kinetics of for families that are their own reverse
(e.g. H_Abstraction) without the thermodynamics of the species involved.
Since we don't have that information within the KineticsFamily, we will
defer the kinetics estimation to a separate step.

Hopefully the kinetics estimation step can be made embarrassingly parallel.