Generate and store reaction template in KineticsFamily.generateReactions().
We will need this information in order to estimate the kinetics using group
additivity values or rate rules. It is difficult to do this at the actual
getKinetics() call because we discard the information about the labeled
atoms along the way. The matched reaction template will probably also be
generally useful, so I've made it a formal attribute of the
TemplateReaction class.