Added kinetics generation step after thermo generation in RMG jobs.
A new method CoreEdgeReactionModel.generateKinetics() was added for this
purpose. Most of this code was in KineticsFamily.generateReactions() until
a few commits ago. The hard part is determining the kinetics for reactions
from a family that is its own reverse (e.g. H_Abstraction). The decision
is to prefer kinetics that are not estimates (via rules or group
additivity) or that are estimates for the exothermic direction.

Hopefully this is something we can parallelize easily.