Fixed bug in how pdep path reactions are handled in processNewReactions().
The problem was that path reactions were only being passed to
addReactionToUnimolecularNetworks() if they were new reactions. This does
not work with the activated species algorithm of Matheu, for which each
path reaction could be placed in multiple partial networks and therefore
needs to be processed every time it occurs. (The
addReactionToUnimolecularNetworks() method already handled repeats
properly, so only a few changes in processNewReactions() were needed.

Closes #52.