PDepReactions in core now converted to reversible reactions.
The idea here is that there are certain circumstances that can cause a
PDepReaction to get into the core without its reverse, e.g. well-skipping
over a well that is itself not in the core. Instead of forcing exploration
of that well (which would lead to much larger networks), we instead simply
discard the reverse PDepReactions - in the core *only* - and mark the
remaining ones as reversible.

It would be better to do this before printing the summary of the model
enlargement, but that would take more work.