New way of choosing forward/reverse reaction estimates for own-reverse families.
Choosing in which direction to estimate the kinetics of H_Abstraction reactions 
is a subtle art. This commit adds a new consideration (is one direction using
a group fitted to only one reaction?) which should reduce the errors associated
with .O-O. radical abstractions (very common) being treated as abstractions 
from .O-OH (which have very little data).  It also adds a commentary to the 
kinetics's comment field justifying the choice.

In the future it would be nice to estimate the uncertainties, and use that to 
make the decision.