Profiling postprocessing just appends to log instead of overwriting.
If you call rmg.py with a capital -P flag ('rmg.py input.py -P') then it will 
just postprocess the profiling data (useful when your '-p' job failed).
It used to wipe the RMG.log file - now it just appends to it.
I also added a little header to that section of the log.