Compute densities of states and ground-state energies for product channels.
The user may provide reactions of the form product channel -> isomer with
the kinetics defined in that direction. In general I think it's better to
apply ILT in the direction the kinetics are known. To enable this, we need
to allow for densities of states and ground-state energies to be computed
for product channels if their states data is available. Note that product
channels could be either unimolecular or bimolecular, and we need to handle
both cases.