Fix bug in E0 estimation within Species.generateThermoData().
There was a path through this method that would result in E0 not being set,
and therefore assuming its initial value of 0. This would happen if the
thermo was obtained in MultiNASA format from a library; currently this
could only occur from GRI-Mech, but others could cause this in the future.
Now, the generated thermo model is always converted to a Wilhoit and the
E0 estimated by extrapolation to ~0 K before any return paths are
encountered.

This addresses almost all of the 10^14 error described in issue #58, and
should make models generated with pressure dependence on much more
relevant.

Closes #58.