Raise exception if PDepReactions are not thermodynamically consist.
Use of the activated species algorithm of Matheu causes a PDepReaction to
be generated as a pair of irreversible forward and reverse reactions, each
coming from a different partial network. There is no guarantee that these
two PDepReaction objects satisfy thermodynamic consistency, although they
should be close. This commit causes a PressureDependenceError to be raised
if a pair of PDepReactions are more than a factor of two away from thermo
consistency (as evaluated at the arbitrarily-chosen conditions of 1000 K
and 1 bar).