Fix bug in graph isomorphism isMappingValid checking.
This bug was preventing O=O from matching O=O in an isMoleculeForbidden call.
The initial mapping is made with the star-labelled atoms, then the mapping 
is checked to see if it is valid. The first copy/paste error meant that 
selfHasEdge==True but otherHasEdge==False, so it returns False. Once that 
is fixed, the second copy/paste bug is needed so that edge2 is 
taken from the right object.

Get this commit and run the new graphTest.py, it will fail.
Re-cythonize your new graph module, and the graphTest will pass.