Also estimate spin multiplicity when generating frequencies.
The spin multiplicity is currently estimated as the number of radical
electrons in the molecule plus one, which is fine for closed-shell species
and monoradicals, but not quite correct for higher-order radicals. The
spin multiplicity acts as the electronic partition function, and so
provides multiplicative factors that influence the generated k(T,P) values.
Therefore it is important to have them as correct as possible.

Closes #57.