Refactor updateConfigurations() to keep bimolecular core species as reactants.
Besides adding comments (and a better docstring) for readability, I've
changed the algorithm slightly so that any configurations consisting
of all core species are treated as reactant channels (if bimolecular).
This is needed so that the partial networks consider both the forward and
reverse directions for these configurations. Otherwise we won't satisfy
thermodynamic consistence because one direction will be a (small but
significant) overestimate.

Closes #58 (again).