Changed thermo consistency exception to warning.
This was done because I'm not convinced that thermo consistency is always
possible using the activated species algorithm to explore pressure-
dependent reaction networks. Fortunately it still appears to be somewhat
rare.

Since we are no longer aborting, we need a heuristic to decide which of
the two possible reactions to keep. For now the heuristic is to keep the
one from the more explored network. If the networks are equally explored,
keep the one that predicts kf(T,P) and kr(T,P) as the faster pair. (Note
that, since kf/kr = Keq, either both are faster or both are slower.)