Fixed numerical bug in calculateMicrocanonicalRateCoefficient().
The multiplication required to get kr(E) from kf(E) involves several terms
of potentially very small numbers. If you're not careful about the order
that you multiply in, you could end up with some unexpected zeros in your
k(E) arrays. This was one of the sources of NaN in Issue #60. This new
implementation should eliminate (or at least significantly reduce) the
times that this occurs. More importantly, I have added a check for this
behavior and now cause an exception to be raised if it still occurs, so
that we can take more action if that does indeed happen.