A few more tweaks to the k(E) generation process.
For association reactions, the forward k(E) is actually stored as
k(E) * b(E), the rate times the Boltzmann distribution for the reactants.
(This allows us to avoid generating frequencies for certain product
channels.) For these reactions the k(E) is computed first, then converted
to k(E) * b(E). In the spirit of making this more numerically robust, I've
changed the order of multiplication to hopefully make it less likely to
walk past the end of the range of a double-precision number.

I've also changed the ILT implementation to use the simple case (nominally
for n = 0) for n very close to zero, as the more general case for n > 0
involves a more expensive convolution and may not be very accurate for,
say, n = 1e-10 (which seems to appear once in a while).