Added uncertainty estimation to Arrhenius.fitToData().
Note that, if you were to propagate the uncertainty in each of the
individual Arrhenius parameters, you would get a larger overall uncertainty
in k(T) than simply by evaluating the sum-of-square error at the optimum
point. This is inherent to the linear least-squares method I believe. (It's
certainly better for this to be an overestimate than an underestimate!)