Make sure k(E) is zero for energies where reactant and product rho(E) is zero.
Clearly we cannot have any reaction between grains where one or both of the
reactant and product has zero Boltzmann population. (This can occur for
certain endothermic barrierless reactions, for which the k(E) at the first
energy above the barrier is nonzero, but the rho(E) for the products is
zero.) This commit adds code that enforces this physical constraint. (If
we did not, we could end up with k(T,P) values that don't satisfy
macroscopic equilibrium, especially at low temperatures where this first
reactive grain is particularly significant.)