Use a simpler heuristic to automatically choose Emax in MEASURE.
One reviewer of my manuscript pointed out that he thought simply choosing
Emax to be many kB*T above the highest energy on the PES would be good
enough. I have come to agree with this, insofar as the current approach
isn't worth the extra time. The actual number of kB*T to use is up for
debate; the small sample size of networks I tested suggest that 40 is a
reasonable place to start.

This will be more important shortly, when we will need to make more calls
to Network.autoGenerateEnergyGrains().