Use interpolation when applying ILT method.
It is often the case that Ea doesn't align well with the grain size,
meaning that we lose some accuracy due to our numerical discretization.
To minimize this, we now interpolate the density of states in the numerator
of the ILT equation to get the approximate value between the grain. (This
interpolation is done semilogarithmically.)

I have seen almost factor-of-two improvements in the k(E) values as a
result of this change, so it's clearly worth doing.