Reaction library is now added to model edge, along with the option of attaching all unused reactions to the output.
The condition file now takes reaction libraries as a list of tuples - i.e.: 
     reactionLibraries = [('GRI-Mech3.0',True),('Glarborg',False)]
If the option on the reaction library is True, then the unused reactions
will be tacked onto the output file as well as chemkin file during each 
save.  However, they will not be included in the core species or reactions.
They will also be tagged as unused by RMG.