Added preliminary support for reading Chemkin files.
The current algorithm is likely biased towards the Chemkin files that
RMG-Java and RMG-Py generate, and may not work for every Chemkin file in
existence. There are also a few assumptions it makes at the moment,
particularly about units. Nonetheless, it is able to parse the Chemkin
files that are saved by RMG-Py, so it's a good start.