Added saveSpeciesDictionary() function to RMG Chemkin output.
This function saves a text file containing the adjacency lists of all
core species to the file chemkin/species_dictionary.txt. (The equivalent
output file in RMG-Java is RMG_Dictionary.txt.) This file is a useful
supplement to the generated Chemkin files, which do not have molecular
structure information associated with them. Note that we don't need to
save one of these with each Chemkin file, as later iterations will always
have at least the same core species as earlier ones.