Added first version of script for comparing two RMG models.
The script diffModels.py loads a pair of Chemkin files and corresponding
species dictionary files, and then compares the models contained within.
In particular, the script currently searches for the reactions that the
two models have in common and creates a parity plot of the k(T) and/or
k(T,P) values in each model.

The script is still maturing, and makes a few assumptions about the two
models (e.g. each reaction only occurs zero or one time in each model),
but it's in a good enough state that I think others might find it useful.