getKineticsForTemplate() now allows choice of estimation method.
The choices are 'group additivity' and 'rate rules'. The latter is the
RMG-Java way of using the rate rules to get the kinetics, which has not
yet been implemented. The group additivity functionality was moved to a
new method estimateKineticsUsingGroupAdditivity(). A placeholder method
estimateKineticsUsingRateRules() was also added, which currently raises
a NotImplementedError.