First version of rate rule-based kinetics estimation (like RMG-Java).
The group additivity-based kinetics estimation scheme is a bit too
immature for use in RMG at the present time. Thus we need to implement
the old-style RMG rate rule averaging scheme (or something reasonably
similar) so we can start using RMG-Py for real systems.

This commit provides the first attempt at doing so, using Jing's thesis
(p. 153-154) as a guide. Note that this implementation does not have the
crazy averaging often seen in RMG-Java yet, simply because this version
only averages *up* the tree. At the moment, siblings with the same
"distance" from the exact node of interest are not averaged; instead, this
implementation gives preference to a more even falling up the tree over
falling up one part of the tree disproportionately to another.