Now averaging all kinetics rate rule estimates within a given distance.
The "distance" of an estimate is the total number of moves towards parents
for each node required to find one or more kinetics estimates. There's no
way a priori to know whether moving up one subtree is "better" or "worse"
than another subtree, so we simply treat each move as equal. I *think*
this is what RMG-Java does.