Don't consider other labeled atoms when descending tree.
When descending the tree, we first match the labeled atoms in the group
with the corresponding atoms in the structure. However, previously we
were allowing non-labeled atoms in the group to match other labeled atoms
in the structure. This led to multiple siblings matching the structure,
which RMG-Py would handle by stopping its tree descent at that point.
By making this change, most of the multiply-matched sibling warnings go
away.

I'm not sure this is the best implementation (it's a bit of a hack), but
it seems to work.