Added kineticsEstimator option to RMG input files.
This gives end users the opportunity to choose between the "rate rules"
and "group additivity" methods. The former is a semi-decent approximation
of the RMG-Java kinetics estimation scheme, while the latter is intended
to be its replacement. Since I'm not sure the group additivity approach
is mature enough for use at the current time - and since more options are
better than fewer - I've left it to the end user to choose.

Note that there are several places where the "default" method is set, so
if you want to make the rate rules method the default, you will have to
change it in multiple spots.

N.B. The "rate rules" approach will not give identical results to RMG-Java,
as it would take more time than it's worth to accomplish this. Instead, it
is more of a facsimile. Certainly the "rate rules" method will give
kinetics more similar to the RMG-Java estimates than the group additivity
method. I cam't speculate as to whether the RMG-Py or RMG-Java
implementation of the "rate rules" method is better.