Can now store termination criteria on reaction system objects.
This is done so that we can allow for specifying termination criteria that
varies between reaction systems. It also makes the code a bit easier to
read, so it seems like a natural thing to do, especially since we define
the TerminationConversion and TerminationTime classes in rmgpy.solver.base
anyway.