Added generateFluxDiagram.py script for generating flux diagrams.
The script accepts as input the path to an RMG-Py input file for a
completed -- or interrupted -- job. It reads the reaction model from the
last Chemkin file and species dictionary, and the reaction systems from the
RMG-Py input file. Each reaction system is simulated using the loaded
reaction model, and a flux diagram is generated. (You will need mencoder
to generate the movie itself.) If species drawings are available, they are
used for the nodes.

The fluxes displayed are actual molecular/molar fluxes (not carbon or heavy
atom fluxes). The function getFluxPairs() determines the pairings of
reactants and products for the purposes of the flux diagram; it probably
needs a better algorithm for handling A + B -> C + D reactions.