Gathered options for flux diagram generation at top of script.
I expect that end users will want to tweak these depending on which
reaction model and reaction systems they are using, so I've tried to make
it easy to find them. It might be worth making them configurable via the
command line, but that's a separate commit. More options can easily be
added if we find them necessary and/or useful.