Added partial method for reading RMG-Java condition file to RMG class.
N.B. The current implementation of the new loadRMGJavaInput() method is
quite incomplete; it only loads the information needed to generate flux
diagrams from RMG-Java models. Someone will need to finish this method in
the future if we want to be able to convert RMG-Java input files to RMG-Py
automatically.

Also needs more comments and debugging!