Reaction.generatePairs() can now handle more than two reactants/products.
Occasionally you will find a reaction of the form A + B -> C + D + E in
a reaction library. We need to be able to at least hazard a guess as to
the flux pairs for this. The new algorithm will pair off reactants and
products by heavy atom count until only a single reactant or product
remains, and then pair the remaining products/reactants off with that
reactant/product. Probably not correct in a bunch of cases, but better
than nothing.