Improvements to the reaction comments saved to Chemkin files.
The objective here is to provide at least a subset of the comment that we
can more easily parse when reading in a Chemkin file (e.g. when generating
flux diagrams). We now have individual lines for the reaction type and
source, and the flux pairs. More such lines can easily be added in the
future.

Also cleaned up a few things in the comments from the kinetics itself.