Updated Chemkin file reader to parse new style of reaction comments.
Besides being cleaner and easier for humans to read, the new-style reaction
comments used in saving Chemkin files are also easier for the computer to
parse. Now we can easily determine the type of reaction (template, library,
or pdep), the flux pairs, and anything else we may want to add in the
future.

This should also cope with RMG-Java-generated Chemkin files, though we
will want to add this style of comments there so we can use that
information, e.g. when diffing two models.