Added directory to species() and transitionState() CanTherm input blocks.
This allows the user to specify a directory on disk to look for the
Gaussian input files for a species or transition state. This is useful if
your Gaussian files are not kept next to your input files. For example, I
use this to refer to the reactant and product species in a CanTherm input
file used to generate kinetics.