Can now save a Diff html page which can compare 2 RMG models using their chemkin and RMG dictionary files.
The Diff output html shows common species and unique species for each model, as well as common reactions and unique reactions, and identifies 
if the kinetics are identical.

Currently, kinetics found in reverse are labeled with "This reaction
was found in reverse."  And the kinetics have not been adjusted for the 
direction shown on the html page.

Checkboxes don't yet work fully for the families/comments/etc.  
Checking them affects both models' lists of unique reactions, 
but not the common reactions.  Current checkboxing categorization
will not work as well for the common reactions because reactions
may not be originating from the same family in both models.