Updated molecule drawing to render output on white background.
It would be better to use a clear background, but Cairo seems to not like
this as much. In particular, the issue comes when we try to clear the
background behind an atom label, for which the resulting output becomes
pixelated and ugly. If we clear to a white background instead the issue
goes away. Since the vast majority of use cases will want a white
background anyway, I'm making this the default.

There may still be issues with SVG output, but PDF should be working much
better now.