When parsing Chemkin files, also handle reactions without RMG-style comments.
The RMG-style comments provide useful information (e.g. when diffing two
models), but not all Chemkin files we would want to try to load will have
these comments. Since we still want to be able to use them, I've added the
necessary logic to handle that case as well. I would have added it earlier,
but I thought it was going to be more difficult than it turned out to be.