Add termination time to systems if necessary during flux diagram generation.
It is possible that you could try to submit a mechanism to the flux diagram
generator with termination conversions that are physically impossible for
that mechanism (due to it missing reactions or simply being
thermodynamically forbidden). Without a termination time the solver would
try to integrate forever in search of this impossible condition. Instead,
we add an explicit (and very large) termination time so that the solver
doesn't get trapped in an infinite loop.