Added RMG-Java chemkin parsing capability back into Py chemkin reader. Also changed method of dealing with duplicate reactions.
For duplicate reactions, if reactions are marked as duplicate, they are 
combined into a single MultiKinetics object ONLY if they are from the
same reaction library.  If they are not, then they do not get combined into one object even if they must be marked as duplicate for chemkin usage.

However, if both reactions are pressure dependent or pressure-independent, and the reactions were not marked as duplicate, then this
raises an error because it will cause chemkin to fail.