Added generateReactions.py for generating reactions from a set of species.
This provides equivalent functionality to the PopulateReactions class in
RMG-Java. You provide an input file similar to regular RMG-Py jobs, but
without the reaction system and solver parameters (since we don't need to
do any ODE solving). The generated HTML and Chemkin output contain all
reactions, including both the core and edge (as we make no distinction
between these for the purposes of the generateReaction script).