Added Network.mapDensitiesOfStates() method.
This method encapsulates the mapping of density of states data from one
set of energies to another. We want to call this code from multiple places,
so it makes sense to make a method out of it.

This is a key function in enabling use of different energies at different
temperatures, which is a nice feature that MEASURE has but FAME does not.