Major refactoring of the Network class.
The primary motivation here is to enable the Network class to store
information about the pressure dependence calculation as it occurs, so
that we can use it for more detailed analyses (e.g. for papers), as well as
minimize code duplication. This is achieved by storing all of the
parameters for the pressure dependence calculation as attributes of the
Network object.

To use the new Network object manually, you first call initialize() to
compute a few parameters that are not dependent on the current conditions.
You then call setConditions() to set the current conditions; all necessary
parameters are then updated as needed to reflect the new conditions.
When you are finished, call clear() to free the memory used by the
parameters.