First version of KineticsFamily.fillKineticsRulesByAveragingUp().
This is designed to mimic a similar function in RMG-Java that attempts to
fill in gaps in the kinetics rate rules for each family by averaging what
rate rules already exist. In Jing Song's thesis she states that, without
this, you would fall up to the top-level nodes most of the time. It seems
that crazy averaging tends to give better estimates than this.

The end goal is a near-exact replication of the RMG-Java kinetics
estimation scheme, because I'm finding a significant number of cases where
it beats the group additivity scheme we are currently using.