Removed KineticsFamily.__getKineticsForRateRule() method.
This was an attempt to mimic the functionality now provided by
fillKineticsRulesByAveragingUp() at the point when the kinetics are needed
instead of when the database is loaded. The idea was to not have to add
the average rate rules to the database, since there are circumstances
under which we don't want them. However, I think it's more efficient in
the long run to fill in the rate rules right away. We can take other steps
to ensure that these are not saved with the database.