Some work on speeding up fillKineticsRulesByAveragingUp().
The current implementation is recursive in such a way as to cause each
node in the combined tree to be visited multiple times. These visits are
not cheap, so I've added a dict to cache the results so we only need to
run the full function once per node. Interestingly, I don't see such a
thing in the RMG-Java implementation, yet it is much faster than the
RMG-Py version, even after factoring in the speed difference between
compiled-vs.-interpreted code.