Replace rate rules with rank 0 with averaged rate rules if possible.
Presumably, if the entry has rank of zero, then we have so little faith in
it that we'd rather use an averaged value if possible. Note that we can't
just overwrite the existing rule, or we won't be able to save it again.
Thus there may yet be work to do to get RMG-Py to prefer the averaged
rule over the rank-0 rule, but at least both are available.

All averaged rate rules are now added with a rank of 10 to make them
distinct from the explicit rate rules.