KineticsFamily.getRateRule() has better handling of multiple matches.
If we have multiple rate rules for a given template, we use the rank and
the index as the deciding factors. The rule with the lowest nonzero rank
that occurs soonest in the library is the one that is used.

Note also that, for R_Recombination, we must remember to check for rules
in both the A + B and B + A combinations, since A and B are from the same
tree. This unfortunately requires a bit of hardcoding of reaction families.