A few more tweaks to get the rate rule scheme matching that of RMG-Java.
* Don't descend past nodes that have an exact rate rule, even if you could
  fill in more of the tree by doing so.

* Only use permutations of the children of all nodes in the template. For
  a family with two subtrees, these are all the distance = 2 combinations.
  Note in particular that we do not use any of the distance = 1 pairs,
  even though they are "closer" to the node we are trying to get an
  averaged value for. However, if the node is a leaf than we do use it
  (since it doesn't have any children otherwise).

These are not necessarily the best course of action to take, but they do
seem to mimic very closely the behavior of RMG-Java's kinetics estimation
algorithm. The only differences that I see left are because RMG-Py uses
dHrxn(0 K) instead of dHrxn(298 K) to fix the barrier heights. These are
small effects (well within the uncertainty of our kinetics estimates),
so they are probably okay.