Better handling of "Others-" nodes when descending tree.
These are special cases for nodes that are intended to be "anything that
doesn't match one of my siblings". My opinion is that we should just be
stopping at the parent node in this case, but for now we need to be able
to match the "Others-" nodes in order to be consistent with the current
database implementation (e.g. match RMG-Java kinetics estimation
procedure). This commit checks for an "Others-" node and descends to it
if appropriate.

Note that RMG-Java calls such nodes "DummyLeaf".