Now filling in averaged rate rules in RMG jobs using rate rules method.
This should now give k(T) estimates that are very close to those of
RMG-Java, with the only significant difference being the effect of using
dHrxn(0 K) instead of dHrxn(298 K) to fix barrier heights. If the quality
of the kinetics estimates in RMG-Py was a reason that people were reluctant
to use Py instead of Java, then it should no longer be an issue.