Handle case where rate rules with rank of zero are returned from depository.
In a previous implementation of looking up kinetics in the depository, I
had assumed that we would never see rules with a rank of zero (since they
would be overwritten with averaged values). However, this turned out to be
a poor assumption. The key here is that we want to keep the rule with the
lowest nonzero rank if possible; if all ranks are zero, then we can keep
the first of them.