Tweaks to choosing the kinetics for families that are their own reverse.
If both directions match exact rate rules, keep the direction with the
lower (but nonzero) rank. Not sure if RMG-Java does this.

Changed a few of the outputted comments to reflect that the k(T) estimates
are not necessarily from group additivity (could also be rate rules).

Fixed a bug in the storing of the template when keeping the reverse
reaction (was showing the template for the forward direction).

Now printing the enthalpy of reaction at 298 K in the kinetics comments.