Improved calculation of axis symmetry numbers.
For the purposes of symmetry number calculation, an axis is actually any
set of adjacent double bonds, including isolated double bonds. Before I
was ignoring isolated double bonds, which meant that the axis symmetry
number was incorrect for molecules such as ethylene.

I've also made an effort to improve the handing of the fragments on either
end of the axis to handle the possibility of radicals on the terminal atoms
in the axis. If there are some ligands and radicals on the terminal atoms,
then the radicals must be considered as an empty ligand for the purposes of
evaluating the symmetry.

Also added a few more unit tests to check that these changes are working as
intended.