In diffModels script, common species and reactions given as lists, not dicts.
This is to preserve the order in which the reactions appear. Instead of a
dict of key-value pairs, the common species and reactions are stored as
lists of 2-tuples. (Incidentally, if you passed this to the dict()
constructor, you would recover a dict in the same form as it was before
this change.)

Also updated saveDiffHTML() to reflect this change.