Better handling of species images path specification in generateFluxDiagram.py.
The previous implementation made some inappropriate assumptions about your
working directory while the script was running. This was especially not
true when using the website to generate your flux diagram movies. The
result was that you would not get the nice pictures of the species in your
movies, and would instead see just the species identifiers.

If a species images path is not given, it tries to use the same place it
previously assumed before falling back to using the species identifiers.