Better handling of energies from scan jobs in Cantherm.
The loadScanEnergies() method now also returns the corresponding angles,
although the angles are still assumed to be equally spaced from 0 to 2pi,
rather than loaded from the log file. Also, if for some reason the user
ran a scan job with opt+freq instead of just opt, the loadScanEnergies()
method will discard the energy read from the freq part of the calculation.