Improved generation of hindered rotor parameters in Cantherm.
The new algorithm will fit both a simple cosine function and a Fourier
series to the potential energy profile. If the cosine function is a good
enough fit, then it will use that model for the rotor. Otherwise it will
use the Fourier series potential as before. The motivation is that I was
observing the cosine potential to actually give a slightly better fit in
a few cases. Furthermore, the cosine potential is much cheaper to use as
a hindered rotor model, so we'd like to use it where we can.

Also improved the plots of the hindered rotor potentials to show the domain
from 0 to 2pi and to enable comparison of both the cosine potential and
Fourier potential fits. The one that was kept is shown with a solid line,
the other with a dashed line.