Fixed bug in generateReactions involving duplicated species/reactions in output.
The bug was that the "core" reactions and species were being added to the
output twice, due to a misunderstanding about how the outputSpeciesList and
outputReactionList lists worked. Since these lists are added to the output
in addition to the core species and reactions, we only need to add the
edge (and not the core) species and reactions to them before calling the
output HTML and Chemkin functions.