CoreEdgeReactionModel.enlarge() can now accept a list of objects to enlarge.
There are certain steps in the model enlargement that only need to be done
after all of the objects being enlarged (by adding to core or exploring in
networks) have been processed. Therefore it makes sense to allow for
enlarging multiple objects at a time. This commit provides such
functionality.