Some more refactoring of CoreEdgeReactionModel.enlarge().
Now generates thermo and kinetics after processing each item in the list
of objects to enlarge, in preparation for parallelization of those aspects
of the process. Also cleaned up the printing of the enlarge summary to
occur after all objects are processed in enlarge(). The key seems to be
that self.newSpeciesList and self.newReactionList must be used as before,
i.e. reset to empty between the objects.