addKineticsRulesFromTrainingSet() handles reactions in either direction.
It is useful to be able to add entries to the training set for reactions
in either the forward or reverse direction of the reaction family. In the
latter case, we need to get the reverse k(T) in order to construct our
rate rule. To do this we need the thermo, which means we need to pass the
thermo database to this function. We also need to compute the reverse
degeneracy.